src/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Document ingestion pipeline with OCR and vector storage"
requires-python = ">=3.11"
dependencies = [
    "common",
    "pdf2image>=1.16",
    "Pillow>=10.0",
    "openai>=1.0",
    "chromadb>=0.5",
    "tiktoken>=0.5",
    "langchain-text-splitters>=0.2",
    "click>=8.0",
]

[project.scripts]
ingestion = "ingestion.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestion"]

[tool.uv.sources]
common = { workspace = true }
